Makefile: use sharedir and PREFIX consistently (#138)
Mixing raw `/usr` and PREFIX means that setting PREFIX (and DESTDIR)
doesn't do what's expected. Instead, use the variables consistently so
that building with a different PREFIX and DESTDIR is possible.

I also put the installation of metadata assets into its own target, so
that it can be invoked without attempting to build and install the
application itself.